"Normalized" user creation method
In a normal AD, the object gets created with the CN (Common Name) and
sAMAccountName different from each other. With the current process,
the function creates the user with the given sAMAccountName and puts it
as CN too, which makes it not editable later in the script since AD blocks
you from editing the CN of a "opened" object.

Example:
CN = "Walker Sean"
sAMAccountName = "seanwlk"